function that prints all the elements of a listint_t list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -6,7 +6,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	int nodes = 0;
+	size_t nodes = 0;
 	
 	if (h == NULL)
 		return (nodes);
